Change alert model

+++ FindMyGolda/app/src/main/java/com/example/findmygolda/alerts/AlertManager.kt
@@ -92,7 +92,7 @@ class AlertManager(val context: Context):ILocationChanged {
                 val alertId = dataSource.insert(
                     Alert(title = branch.name,
                         description = branch.discounts,
-                        branchId = branch.id.toInt(),
+                        branchId = branch.id,
                         isRead = false)
                 )
                 notify(branch.name, branch.discounts, alertId, NOTIFICATION_IMAGE_ICON)
